requirements-dev.txt
asgiref==3.8.1
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
Django==4.2.16
django-admin-logs==1.3.0
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
mysqlclient==2.2.4
packaging==24.1
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.2.1
rsa==4.9
six==1.16.0
spark-parser==1.8.9
sqlparse==0.5.1
uncompyle6==3.9.2
whitenoise==6.7.0
xdis==6.1.1
